Move the discord
Move the discord stuff into the discordcontroller section for SoC
